+++ wordle-web/layouts/default.vue
@@ -1,0 +1,43 @@
+<template>
+  <v-app>
+    <v-app-bar color="primary" :elevation="2">
+      <template v-slot:prepend>
+        <v-icon color="secondary"> mdi-book </v-icon>
+      </template>
+
+      <v-app-bar-title>WordleApp</v-app-bar-title>
+
+      <v-btn to="/">Home</v-btn>
+      <v-btn icon="mdi-theme-light-dark" @click="toggleTheme" />
+      <v-btn icon="mdi-help-circle" @click="showHelpDialog = true" />
+      <HelpDialog v-model="showHelpDialog" />
+    </v-app-bar>
+    <v-main>
+      <slot />
+    </v-main>
+  </v-app>
+</template>
+
+<script setup lang="ts">
+import { useTheme } from "vuetify";
+import nuxtStorage from "nuxt-storage";
+
+const router = useRouter();
+const theme = useTheme();
+const showHelpDialog = ref(false);
+
+function toggleTheme() {
+  if (theme.global.name.value === "light") {
+    theme.global.name.value = "dark";
+  } else {
+    theme.global.name.value = "light";
+  }
+
+  nuxtStorage.localStorage.setData("theme", theme.global.name.value);
+}
+
+onMounted(() => {
+  var defaultTheme = nuxtStorage.localStorage.getData("theme");
+  theme.global.name.value = defaultTheme ?? "dark";
+});
+</script>
